Add a relay module to control a fan
The fan relay module has been added with it's signal pin attached to pin 7.

I've added three more operating parameters and I've commented them all for future reference:

	#define FAN_ON_TEMP_C      26        // Temperatures above this value will switch the fan on.
	#define FAN_ON_HUMIDITY_PC 50        // Humidity levels above this value will switch the fan on.
	#define MIN_FAN_ON_TIME_MS 10000     // The minimum duration the fan should stay on for (prevents thrashing the relay when the conditions waver between on and off values.

relayModule has been renamed to WaterRelayModule and there's a FanRelayModule now to make things more obvious.

I removed the menu stuff because that's not doing anything yet. With more thought I'm considering making one of the buttons a PAUSE button so that the system can be halted if it's going nuts. I guess switching off the power would work as well. hehe.

I've added the "F" macro on the fixed strings to save RAM.

I've added the DHT readings to the diagnostics so you'll see the temperature and humidity shown there as well.

I put all of the fan control logic in a function called FanControl. There you'll find the logic that controls how long the fan stays on for once switched on. Without a real time clock (RTC) we're just using the millis function to ask the Arduino how long it's been (in milliseconds) since it was powered on. Because this overflows back to zero after approximately 50 days, there's logic in there that will check the temperature and switch the fan on/off if the time mark for when the fan was switched on is zero (which it is at program start and whenever we turn it off) and if the millis now is after the time it was turned on plus the minimum on time, OR.. if the millis now is less than the fan on time (i.e. it overflowed to zero since we turned the fan on). All that's saying is that there can be times when the fan is on for less than the minimum on time. This can happen if the fan is turned on just before the millis overflows to zero. As soon as it overflows to zero we check the fan again and reset the on time.